Remove the need for an allocator
There is not a root source file for tests yet
(https://github.com/ziglang/zig/issues/6621), so it's impossible to pass
the logging init function an allocator.

A hacky custom type was used to allow std.debug.printSourceAtAddress to
print to this type so it is possible to extract the file name and line
number.

This is all a huge workaround until zig gets better at this stuff.